veo_agent/requirements.txt
fastapi
uvicorn[standard]
pydantic
google-generativeai